+++ script/deploy_static.sh
@@ -17,6 +17,6 @@ git add -A
 git commit -m 'deploy'
 
 # if you are deploying to https://<USERNAME>.github.io/<REPO>
-git push -f git@github.com:fabienwnklr/free-drawing-demo.git master
+git push -f https://github.com/fabienwnklr/free-drawing-demo.git master
 
 cd -