+++ src/components/Toast/Toast.ts
@@ -1,16 +1,44 @@
 import type { Drawer } from '@/Drawer';
 import './toast.scss';
+import MicroEvent from '@/utils/MicroEvent';
+import { AvailableTypes } from '@/@types/toast';
+import InfoIcon from '@/icons/info.svg?raw';
+import WarningIcon from '@/icons/warning.svg?raw';
+import ErrorIcon from '@/icons/error.svg?raw';
 
-export class Toast {
+export class Toast extends MicroEvent {
   $toastContainer: HTMLDivElement;
-  timeout: number = 3000;
+  timeout: number = 5000;
 
-  constructor(drawer: Drawer, message: string, type: string = 'neutral', timeout: number = 3000) {
+  constructor(
+    drawer: Drawer,
+    message: string,
+    type: AvailableTypes = 'neutral',
+    timeout: number = 5000,
+    showIcon = true
+  ) {
+    super();
+
     this.timeout = timeout;
     this.$toastContainer = document.createElement('div');
     this.$toastContainer.classList.add('drawer-toast', type);
 
-    this.$toastContainer.innerHTML = message;
+    // default is neutral, so no icon set
+    let icon = '';
+
+    if (type === 'info') {
+      icon = InfoIcon;
+    } else if (type === 'warning') {
+      icon = WarningIcon;
+    } else if (type === 'error' || type === 'danger') {
+      icon = ErrorIcon;
+    }
+
+    this.$toastContainer.innerHTML = icon + message;
+
+    if (!showIcon) {
+      this.$toastContainer.innerHTML = message;
+    }
 
     drawer.$drawerContainer.append(this.$toastContainer);
   }
@@ -21,7 +49,9 @@ export class Toast {
     if (this.timeout) {
       setTimeout(() => {
         this.hide();
-      }, this.timeout)
+
+        this.trigger('fd.toast.hide');
+      }, this.timeout);
     }
   }
 
@@ -29,7 +59,7 @@ export class Toast {
     this.$toastContainer.addEventListener('animationend', () => {
       this.$toastContainer.classList.remove('show', 'hide');
       this.$toastContainer.classList.remove();
-    })
+    });
     this.$toastContainer.classList.add('hide');
   }